+++ docs/.vuepress/config.js
@@ -132,13 +132,6 @@ module.exports = {
           alias: '/api/events',
           collapsable: true,
         },
-        {
-          title: 'Error Types',
-          collapsable: true,
-          children: fs.readdirSync(path.join(__dirname, '../api/classes'))
-            .filter((n) => n.match(/.*error\.md$/))
-            .map(f => `/api/classes/${f}`)
-        },
       ],
       '/': [
         {
@@ -178,6 +171,7 @@ module.exports = {
             ['/guide/integration-with-react', 'Integration with React'],
             ['/guide/integration-with-vue', 'Integration with Vue'],
             ['/guide/integration-with-angular', 'Integration with Angular'],
+            ['/guide/integration-with-svelte', 'Integration with Svelte'],
           ]
         },
         {